Store in in GameState

+++ src/Lynx/Model/GameState.cs
@@ -5,11 +5,11 @@ public readonly struct GameState
 
     public readonly byte Castle;
 
-    public readonly sbyte CapturedPiece;
+    public readonly int CapturedPiece;
 
     public readonly BoardSquare EnPassant;
 
-    public GameState(long zobristKey, sbyte capturedPiece, byte castle, BoardSquare enpassant)
+    public GameState(long zobristKey, int capturedPiece, byte castle, BoardSquare enpassant)
     {
         CapturedPiece = capturedPiece;
         Castle = castle;
